pyproject.toml
[project]
name = "pypredict-mcp"
version = "0.1.0"
description = "MCP server for predicting satellite transits over a given location."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cachetools>=6.1.0",
    "dotenv>=0.9.9",
    "gradio>=5.34.2",
    "mcp>=1.9.4",
    "openai-agents>=0.0.19",
    "pydantic>=2.11.7",
    "pypredict>=1.7.2",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
